+++ aztec/src/main/kotlin/org/wordpress/aztec/spans/AztecListSpan.kt
@@ -1,7 +1,6 @@
 package org.wordpress.aztec.spans
 
 import android.graphics.Paint
-import android.text.Layout
 import android.text.Spanned
 import android.text.style.LeadingMarginSpan
 import android.text.style.LineHeightSpan
@@ -9,12 +8,10 @@ import android.text.style.UpdateLayout
 import org.wordpress.aztec.Constants
 
 abstract class AztecListSpan(override var nestingLevel: Int,
-                             var verticalPadding: Int = 0,
-                             override var align: Layout.Alignment? = null
-    ) : LeadingMarginSpan.Standard(0),
+                             var verticalPadding: Int = 0
+) : LeadingMarginSpan.Standard(0),
         LineHeightSpan,
         UpdateLayout,
-        IAztecAlignmentSpan,
         IAztecBlockSpan {
     override var endBeforeBleed: Int = -1
     override var startBeforeCollapse: Int = -1