+++ aztec/src/test/kotlin/org/wordpress/aztec/CssStyleAttributeTest.kt
@@ -28,7 +28,7 @@ class CssStyleAttributeTest : AndroidTestCase() {
     @Before
     fun init() {
         context = ApplicationProvider.getApplicationContext()
-        parser = AztecParser()
+        parser = AztecParser(AlignmentRendering.SPAN_LEVEL)
     }
 
     @Test